requirements.txt
flask==3.0.0
python-dotenv==1.0.1
supabase==2.6.0